Refactor tokenize function & add comments
The tokenize function required the caller to pre-allocate a buffer to
store the tokens in. Now, the array is allocated dynamically by the
function to only require the minimal amount of heap space.

The buffer & it's size are now returned to the caller in a dedicated
struct. A new function, `matsplat_tokenize_destory`, is available to
free up the heap space and reset the lenght.